chore: move ColorPicker into ui
