+++ apps/frontend/package.json
@@ -21,7 +21,6 @@
     "@react-oauth/google": "^0.11.1",
     "@tremor/react": "^3.9.2",
     "@types/react-color": "^3.0.10",
-    "@types/react-syntax-highlighter": "^15.5.11",
     "@uiw/react-color-sketch": "^2.0.3",
     "graphql": "^16.8.1",
     "js-cookie": "^3.0.5",
@@ -50,7 +49,7 @@
     "@types/js-cookie": "^3.0.5",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
-    "@types/react-syntax-highlighter": "^15.5.10",
+    "@types/react-syntax-highlighter": "^15.5.11",
     "@typescript-eslint/eslint-plugin": "^6.0.0",
     "@typescript-eslint/parser": "^6.0.0",
     "@vitejs/plugin-react-swc": "^3.3.2",
